feat: easysave client interface wpf updated
