--- EasySaveWPF/EasySaveWPFClient/ViewNS/Create.xaml
@@ -1,129 +1,0 @@
-﻿<Page x:Class="EasySaveWPF.ViewNS.Create"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:view="clr-namespace:EasySaveWPF.ViewNS"
-      mc:Ignorable="d" 
-      Height="600" d:DesignWidth="920"
-      Title="Create">
-    <Page.Resources>
-        <Style x:Key="ModernButton" TargetType="Button">
-            <Setter Property="Background" Value="#87CEFA"/>
-            <Setter Property="Foreground" Value="black"/>
-            <Setter Property="BorderThickness" Value="0"/>
-            <Setter Property="Padding" Value="10"/>
-            <Setter Property="Margin" Value="5"/>
-            <Setter Property="Template">
-                <Setter.Value>
-                    <ControlTemplate TargetType="Button">
-                        <Border Background="{TemplateBinding Background}" CornerRadius="10" Width="100" Height="40">
-                            <ContentPresenter HorizontalAlignment="Center" VerticalAlignment="Center"/>
-                        </Border>
-                    </ControlTemplate>
-                </Setter.Value>
-            </Setter>
-            <Style.Triggers>
-                <Trigger Property="IsMouseOver" Value="True">
-                    <Setter Property="Background" Value="#E0E0E0"/>
-                </Trigger>
-                <Trigger Property="IsPressed" Value="True">
-                    <Setter Property="Background" Value="#BDBDBD"/>
-                </Trigger>
-            </Style.Triggers>
-        </Style>
-        <Style TargetType="ComboBox" x:Key="Flat_ComboBox">
-            <Setter Property="HorizontalAlignment" Value="Stretch"/>
-            <Setter Property="VerticalAlignment" Value="Top"/>
-            <Setter Property="MinWidth" Value="60"/>
-            <Setter Property="UIElement.SnapsToDevicePixels" Value="True"/>
-            <Setter Property="ScrollViewer.HorizontalScrollBarVisibility" Value="Disabled"/>
-            <Setter Property="ScrollViewer.VerticalScrollBarVisibility" Value="Auto"/>
-            <Setter Property="ScrollViewer.CanContentScroll" Value="True"/>
-            <Setter Property="TextElement.Foreground" Value="Black"/>
-            <Setter Property="FrameworkElement.FocusVisualStyle" Value="{x:Null}"/>
-            <Setter Property="BorderThickness" Value="0" />
-            <Setter Property="Background" Value="White" />
-
-            <Style.Triggers>
-                <Trigger Property="IsMouseOver" Value="True">
-                    <Setter Property="Background" Value="LightSkyBlue" />
-                </Trigger>
-            </Style.Triggers>
-
-        </Style>
-        <Style TargetType="Button" x:Key="Flat_Button">
-            <Setter Property="Foreground" Value="Black"/>
-            <Setter Property="BorderThickness" Value="1"/>
-            <Setter Property="BorderBrush" Value="LightGray"/>
-            <Setter Property="Padding" Value="5"/>
-            <Setter Property="Margin" Value="5"/>
-            <Setter Property="Template">
-                <Setter.Value>
-                    <ControlTemplate TargetType="Button">
-                        <Border Background="{TemplateBinding Background}"
-                        BorderBrush="{TemplateBinding BorderBrush}"
-                        BorderThickness="{TemplateBinding BorderThickness}">
-                            <ContentPresenter HorizontalAlignment="Center"
-                                      VerticalAlignment="Center"/>
-                        </Border>
-                        <ControlTemplate.Triggers>
-                            <Trigger Property="IsMouseOver" Value="True">
-                                <Setter Property="Background" Value="LightSkyBlue"/>
-                            </Trigger>
-                            <Trigger Property="IsPressed" Value="True">
-                                <Setter Property="Background" Value="LightGray"/>
-                            </Trigger>
-                        </ControlTemplate.Triggers>
-                    </ControlTemplate>
-                </Setter.Value>
-            </Setter>
-        </Style>
-
-    </Page.Resources>
-   
-
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="60"/>
-            <RowDefinition Height="60"/>
-            <RowDefinition Height="60"/>
-            <RowDefinition Height="60"/>
-            <RowDefinition Height="100"/>
-            <RowDefinition Height="260"/>
-        </Grid.RowDefinitions>
-        <TextBlock Grid.Row="0" Text="{DynamicResource ChooseSave}" Foreground="#87CEFA" FontSize="28" HorizontalAlignment="Left" Margin="0,0,0,10"/>
-
-        <!-- select parth source -->
-        <StackPanel Orientation="Horizontal" Margin="20" Grid.Row="1" VerticalAlignment="Center">
-            <TextBox Text="{DynamicResource SelectSourceFolderButton}" BorderThickness="0"  />
-            <Button Style="{StaticResource Flat_Button}" Margin="3" Name="btnOpenFile" Click="OpenFolderSource_Click" Content="..." />
-            <TextBox Text="{Binding OpenFileSrc, Mode=TwoWay}" BorderThickness="0"  />
-        </StackPanel>
-        
-        <!-- select parth target -->
-        <StackPanel Orientation="Horizontal" Margin="20" Grid.Row="2" VerticalAlignment="Center">
-            <TextBox Text="{DynamicResource SelectDestinationFolderButton}" BorderThickness="0"  />
-            <Button Style="{StaticResource Flat_Button}" Margin="3" Name="btnOpenFile2" Click="OpenFolderDest_Click" Content="..." />
-            <TextBox Text="{Binding OpenFileDest, Mode=TwoWay}" BorderThickness="0"  />
-        </StackPanel> 
-
-        <StackPanel Margin="20" Grid.Row="3"  HorizontalAlignment="Left" Orientation="Horizontal">
-            <TextBlock Text="{DynamicResource SaveFormat}" Margin="0,0,5,0"  VerticalAlignment="Center" />
-            <ComboBox x:Name="saveTypeComboBox"
-                Style="{StaticResource Flat_ComboBox}" 
-                      SelectedItem="{Binding SaveType}"
-                      Width="100" Grid.Row="3" 
-                      VerticalAlignment="Center"
-                      SelectionChanged="SaveTypeComboBox_SelectionChanged">
-                <ComboBoxItem Content="Différentiel"/>
-                <ComboBoxItem Content="Complète"/>
-            </ComboBox>
-        </StackPanel>
-
-        <!-- Boutons en bas -->
-        <StackPanel  Margin="20" Grid.Row="4" Orientation="Horizontal" HorizontalAlignment="Right" VerticalAlignment="Bottom" >
-            <Button Style="{DynamicResource ModernButton}" Content="{DynamicResource AddButton}" Name="btnAddSave" Click="btnAddSave_Click" Margin="5"/>
-        </StackPanel>
-    </Grid>
-</Page>
