--- EasySaveWPF/EasySaveWPFClient/ViewNS/Home.xaml
@@ -1,77 +1,0 @@
-﻿<Page x:Class="EasySaveWPF.ViewNS.Home"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:view="clr-namespace:EasySaveWPF.ViewNS"
-      mc:Ignorable="d" 
-      Background="Transparent" 
-      Height="600" d:DesignWidth="920"
-      Title="Home">
-    <Page.Resources>
-        <Style x:Key="ModernButton" TargetType="Button">
-            <Setter Property="Background" Value="#87CEFA"/>
-            <Setter Property="Foreground" Value="black"/>
-            <Setter Property="BorderThickness" Value="0"/>
-            <Setter Property="Padding" Value="10"/>
-            <Setter Property="Margin" Value="5"/>
-            <Setter Property="Template">
-                <Setter.Value>
-                    <ControlTemplate TargetType="Button">
-                        <Border Background="{TemplateBinding Background}" CornerRadius="10" Width="100" Height="40">
-                            <ContentPresenter HorizontalAlignment="Center" VerticalAlignment="Center"/>
-                        </Border>
-                    </ControlTemplate>
-                </Setter.Value>
-            </Setter>
-            <Style.Triggers>
-                <Trigger Property="IsMouseOver" Value="True">
-                    <Setter Property="Background" Value="#E0E0E0"/>
-                </Trigger>
-                <Trigger Property="IsPressed" Value="True">
-                    <Setter Property="Background" Value="#BDBDBD"/>
-                </Trigger>
-            </Style.Triggers>
-        </Style>
-    </Page.Resources>
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="60"/>
-            <RowDefinition Height="240"/>
-            <RowDefinition Height="120"/>
-            <RowDefinition Height="180"/>
-        </Grid.RowDefinitions>
-        <TextBlock Grid.Row="0" 
-                   Text="{DynamicResource AddButton}" 
-                   Foreground="#87CEFA" 
-                   FontSize="28" 
-                   HorizontalAlignment="Left" 
-                   Margin="0,0,0,10"/>
-        <ScrollViewer Grid.Row="1" Margin="100,0,100,0"
-              VerticalScrollBarVisibility="Auto"
-              HorizontalScrollBarVisibility="Auto">
-            <ScrollViewer.Resources>
-                <Style TargetType="ScrollBar" />
-            </ScrollViewer.Resources>
-            <ListBox Name="ItemSelecteds" 
-             SelectionMode="Extended" 
-             ItemsSource="{Binding Items}"
-             Style="{DynamicResource ModernListBox}"/>
-        </ScrollViewer>
-
-
-        <!-- Boutons en bas -->
-        <StackPanel Grid.Row="2" Orientation="Horizontal" HorizontalAlignment="Center" VerticalAlignment="Bottom" Margin="0,0,0,10">
-            <Button Content="{DynamicResource DeleteButton}" 
-                Click="ButtonDelete_Click" 
-                Margin="5"
-                Style="{DynamicResource ModernButton}"/>
-            <Button Content="{DynamicResource ExecuteButton}" 
-                    Click="ButtonExecute_Click"  
-                    Margin="5" 
-                    Command="{Binding ClickCommand}" 
-                    Style="{StaticResource ModernButton}"/>
-        </StackPanel>
-        <TextBox Grid.Row="3" Text="{Binding ErrorText}" BorderThickness="0" Foreground="Red" FontStyle="Italic"/>
-    </Grid>
-</Page>
