--- EasySaveWPF/EasySaveWPFClient/ViewNS/PlayBreak.xaml
@@ -1,23 +1,0 @@
-﻿<Window x:Class="EasySaveWPF.ViewNS.PlayBreak"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:EasySaveWPF.ViewNS"
-        mc:Ignorable="d"
-        Title="Loading" Height="150" Width="400">
-    
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="*" />
-            <ColumnDefinition Width="Auto" />
-        </Grid.ColumnDefinitions>
-        <ScrollViewer Grid.Row="1" Grid.Column="0">
-            <ListBox x:Name="CurrentSave" SelectedItem="{Binding CurrentSaveSelected}" ItemsSource="{Binding CurrentSave}"/>
-        </ScrollViewer>
-        <StackPanel Grid.Column="1" Orientation="Vertical">
-            <Button Content="✖"  Grid.Column="1" Margin="5" Click="Delete_Click" CommandParameter="{Binding ElementName=listBox}"/>
-            <Button Name="Play" Click="Play_Click" Grid.Column="1" Margin="5"/>
-        </StackPanel>
-    </Grid>
-</Window>
